5_RAG_App/requirements.txt
streamlit
openai
chromadb
python-dotenv
